env variable to disable validation, error when profile not supported
Added new environment variable to disable FHIR validation. Added new
functionality to test if all resources are declaring at least on
supported profile. A validation error is raised if no supported profile
is defined. Profiles are supported if they are declared in the root
validation package (currently de.gecco | 1.0.5) or are dependencies of
the StructureDefinitions. Only profile with abstract = false and kind =
resource are supported as claimed profiles by resources beeing
validated.