adds dry-run functionality to data-send process
The data-send process can now be instructed to perform a dry-run only
using the optional input-parameter "dry-run" with a value of "true". A
dry-run will abort the data-send process after validation end enryption
of the transfer bundle but before the bundle is stored in the local DSF
FHIR server and a Task to start the data-translate process is send to
the GECCO Transfer Hub.