added bundle size to GTH output, fixed errors not showing up in output
Changes to the Task resource need to be stored in the leading task
process variable since it is the only one that survives the return task
restarting the process via the intermediate message throw events.

The start-data-translate process Task now shows the size (in bytes) of
the encrypted transfer bundle.