reworked trigger process with unique instances based on BPMN signals
The new trigger process implementation uses an event based gateway to
either wait for a timer to fire after 24 hours or a global signal to
abort.
The abort signal is send as the first step of the trigger process to
stop existing instances and based on a stopDataTrigger Task resource.
The new implementation does not require (and actually does not allow) to
specify a business-key via Task resource.
If a stopDataTrigger Task resource is send while the trigger process is
searching for new/changed data, nothing happens. Stopping the trigger
process is only possible while the process is sleeping but than is
stopped immediately.